pyproject.toml
[tool.poetry]
name = "linked-ipfs"
version = "0.1.0"
description = ""
authors = ["Egemen Göl <egemengol@gmail.com>"]
readme = "README.md"
packages = [{include = "linked_ipfs"}]

[tool.poetry.dependencies]
python = "^3.10"
typer = {extras = ["all"], version = "^0.7.0"}
pydantic = {extras = ["dotenv"], version = "^1.10.4"}
requests = "^2.28.1"
py-cid = "^0.3.0"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
isort = "^5.11.4"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
